+++ src/com/skilldistillery/entities/JetImpl.java
@@ -2,7 +2,7 @@ package com.skilldistillery.entities;
 
 public class JetImpl extends Jet {
 	
-	public JetImpl(String model, int speed, int range, double price) {
+	public JetImpl(String model, int speed, int range, long price) {
 		this.model = model;
 		this.speed=speed;
 		this.range = range;
